Refactors the socket broadcast client
This commit refactors the socket broadcast client. It switches out
the underlying queue to a circular buffer, that way the memory pressure
of the broadcast client remains constant without needing to do work at
sporadic intervals. The new structure should hopefully also make it
easier for someon to reason about the websocket layer.

Signed-off-by: Arjo Chakravarty <arjoc@google.com>