Add Backward-ROS for improved logging in event of segfaults (#327)
* Add Backward-ROS for improved logging in event of segfaults

`backward-ros` provides a clean way to catch unhandled exceptions and
segfaults and is able to print a stack trace when a program crashes.
This PR adds `backward-ros` to rmf ros2 nodes so that we can trace
causes for segfaults.

Note: This will only work for pure C++ nodes. If we want logging in
downstream fleetadapters they will have to handle that themselves either
using `backward-cpp` or `faulthandler` in python.

Signed-off-by: Arjo Chakravarty <arjoc@google.com>

* Alphabetize deps

Signed-off-by: Arjo Chakravarty <arjoc@google.com>

---------

Signed-off-by: Arjo Chakravarty <arjoc@google.com>
Co-authored-by: Yadu <yadunund@openrobotics.org>