Completely Excise Reservation Cancellation pathways
As noticed by @piliwilliam0306 in #411, the cancellation pathway was not being
called. This commit completely excises the reservation cancellation API. The
rationale for this is that in most cases, we should be using the
Finishing Task. In cases we are not, the use of a reservation system is
actually somewhat ill-defined. For instance say a robot is going from
place A->B. Without a finishing task, it is not clear at all whether the
robot should hold A or B or None of them as the robot may be closer to
A, closer to B or just stuck midway.

In the case when a finishing task is assigned, everytime the robot needs
to move (see GoToPlace.cpp) the reservation negotiator is invoked. The
reservation negotiator will check if a previous reservation was held. If
it was, then it will go release any ticket that was previously held. In
the event of a cancellation, the finishing task will simply take over.

Signed-off-by: Arjo Chakravarty <arjoc@google.com>